+++ Coffee Brewer/Views/History/Charts/ChartPreview.swift
@@ -141,7 +141,7 @@ struct ChartPreview: View {
                         let brew = Brew(context: context)
                         brew.id = UUID()
                         brew.date = Date().addingTimeInterval(TimeInterval(-i * 86400))
-                        brew.recipeGrindSize = Int16.random(in: 10...30)
+                        brew.recipeGrindSize = Double.random(in: 10...30)
                         brew.rating = Int16.random(in: 2...5)
                         brew.recipeName = "Sample Recipe \(i)"
                         brew.roasterName = ["Blue Bottle", "Stumptown", "Intelligentsia"].randomElement()!
